feat(theme-builder): move test.ts files into __tests__ folder
